Moving code from bin/util into tikv, and crate docs (#4538)
* tikv: add metadata to Cargo.toml


* readme: minor tweaks


* tikv: add simple crate docs


* tikv: Add simple config module docs


* config: make LAST_CONFIG_FILE private


* More docs


* bin/*: Move code from binaries into tikv lib


* server: Move check_system_config into tikv


* server: move pre_start into tikv


* server: move run_raft_server into tikv


* server: extract run_tikv to tikv::binutil::server


* tikv: remove some broken links from pre_start docs


* docs: Capitalize Titanium, lowercase key-value
