+++ src/binutil/mod.rs
@@ -1,6 +1,7 @@
+pub mod server;
 #[macro_use]
-pub(crate) mod setup;
-pub(crate) mod signal_handler;
+pub mod setup;
+pub mod signal_handler;
 
 /// Returns the tikv version information.
 pub fn tikv_version_info() -> String {
