+++ src/binutil/setup.rs
@@ -1,28 +1,28 @@
 // Copyright 2018 TiKV Project Authors. Licensed under Apache-2.0.
 
 use std::borrow::ToOwned;
-use std::process;
 use std::sync::atomic::{AtomicBool, Ordering};
 
 use chrono;
 use clap::ArgMatches;
 
-use tikv::config::{MetricConfig, TiKvConfig};
+use crate::config::{MetricConfig, TiKvConfig};
 use tikv_util::collections::HashMap;
 use tikv_util::{self, logger};
 
 // A workaround for checking if log is initialized.
 pub static LOG_INITIALIZED: AtomicBool = AtomicBool::new(false);
 
+#[macro_export]
 macro_rules! fatal {
     ($lvl:expr, $($arg:tt)+) => ({
-        if LOG_INITIALIZED.load(Ordering::SeqCst) {
+        if $crate::binutil::setup::LOG_INITIALIZED.load(::std::sync::atomic::Ordering::SeqCst) {
             crit!($lvl, $($arg)+);
         } else {
             eprintln!($lvl, $($arg)+);
         }
         slog_global::clear_global();
-        process::exit(1)
+        ::std::process::exit(1)
     })
 }
