+++ src/binutil/signal_handler.rs
