+++ src/lib.rs
@@ -1,5 +1,21 @@
 // Copyright 2016 TiKV Project Authors. Licensed under Apache-2.0.
 
+//! TiKV - A distributed key/value database
+//!
+//! TiKV ("Ti" stands for Titanium) is an open source distributed
+//! transactional key-value database. Unlike other traditional NoSQL
+//! systems, TiKV not only provides classical key-value APIs, but also
+//! transactional APIs with ACID compliance. TiKV was originally
+//! created to complement [TiDB], a distributed HTAP database
+//! compatible with the MySQL protocol.
+//!
+//! [TiDB]: https://github.com/pingcap/tidb
+//!
+//! The design of TiKV is inspired by some great distributed systems
+//! from Google, such as BigTable, Spanner, and Percolator, and some
+//! of the latest achievements in academia in recent years, such as
+//! the Raft consensus algorithm.
+
 #![crate_type = "lib"]
 #![cfg_attr(test, feature(test))]
 #![recursion_limit = "200"]
@@ -28,6 +44,7 @@ extern crate serde_derive;
     slog_warn,
     slog_info,
     slog_debug,
+    slog_crit,
     slog_log,
     slog_record,
     slog_b,
@@ -49,6 +66,7 @@ extern crate tikv_util;
 #[cfg(test)]
 extern crate test;
 
+pub mod binutil;
 pub mod config;
 pub mod coprocessor;
 pub mod import;
